completed local testing

+++ app/Http/Controllers/PusherController.php
@@ -11,19 +11,25 @@ class PusherController extends Controller
      * Delete the showdown if the only user has left the channel.
      *
      * @param PusherChannelVacatedRequest $request
-     * @return void
+     * @return \Illuminate\Http\JsonResponse
      */
     public function channelVacated(PusherChannelVacatedRequest $request)
     {
-        $showdownId = explode('.', $request->input('channel'))[1];
-        $showdown = Showdown::findOrFail($showdownId);
+        $events = $request->input('events');
 
-        if (!$showdown) {
-            abort(403);
+        for ($i = 0; $i < count($events); $i++) {
+            if ($events[$i]['name'] !== 'channel_vacated') {
+                continue;
+            }
+
+            $showdownId = explode('.', $events[$i]['channel'])[1];
+
+            $showdown = Showdown::find($showdownId);
+            if ($showdown && $showdown->combatants()->count() < 2) {
+                $showdown->delete();
+            }
         }
 
-        if ($showdown->combatants()->count() < 2) {
-            $showdown->delete();
-        }
+        return response()->json(['success' => true], 200);
     }
 }
